feat: add ORDER BY and LIMIT OFFSET methods
